fix(ci): keep context menu guards within budgets
